% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Likelihoods_and_gradients.R
\name{addterm.efp}
\alias{addterm.efp}
\title{Estimate capture probabilites from electrofishing data}
\usage{
addterm.efp(object, scope, scale = 0, test = c("none", "Chisq", "F"),
  k = 2, sorted = FALSE, trace = FALSE, ...)
}
\arguments{
\item{formula}{a formula object}

\item{data}{a data.frame containing all relavent info}
}
\value{
glm type object
}
\description{
This function uses the marginal likelihood of capture probabilities
to estimate model parameters
}
\examples{
# none yet
}

