% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/old-juvdens-functions.R
\name{buildTopo}
\alias{buildTopo}
\title{Title}
\usage{
buildTopo(lines)
}
\arguments{
\item{nb}{an input parameter}
}
\value{
what does it return
}
\description{
Details
}
\details{
Description - This function does stuff.
}
\examples{
y <- rivs[rivs $ CATCH_NAME == "River Irvine",]
g <- buildTopo(y)
plot(g, vertex.label=NA, vertex.size=2,vertex.size2=2)
}
\seealso{
\code{\link{getQMat}} for creating a GMRF from a graph
}

