% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{CLmodel}
\alias{CLmodel}
\title{Modelling Framework for the Estimation of Salmonid Abundance in Scottish Rivers}
\description{
Functions to estimate juvenile densities of Salmon and Trout on river networks.
}
