% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Likelihoods_and_gradients.R
\name{efp.fit}
\alias{efp.fit}
\title{Estimate capture probabilites from electrofishing data}
\usage{
efp.fit(X, y, data, ...)
}
\arguments{
\item{data}{a data.frame containing all relavent info}

\item{formula}{a formula object}
}
\value{
glm type object
}
\description{
This function uses the marginal likelihood of capture probabilities
to estimate model parameters
}
\examples{
# none yet
}

