% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Likelihoods_and_gradients.R
\name{efp}
\alias{efp}
\title{Estimate capture probabilites from electrofishing data}
\usage{
efp(formula, data = NULL, verbose = TRUE, init = "0")
}
\arguments{
\item{formula}{a formula object}

\item{data}{a data.frame containing all relavent info}

\item{formula}{a formula object}

\item{data}{a data.frame containing all relavent info}
}
\value{
glm type object

glm type object
}
\description{
This function uses the marginal likelihood of capture probabilities
to estimate model parameters
}
\examples{
# none yet
Estimate capture probabilites from electrofishing data

This function uses the marginal likelihood of capture probabilities
to estimate model parameters
}

