% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-juvdens-functions.R
\name{getWRW1Mat}
\alias{getWRW1Mat}
\title{Compute Weighted RW1 matrix}
\usage{
getWRW1Mat(g)
}
\arguments{
\item{g}{a graph object}
}
\value{
precision matrix
}
\description{
Details
}
\details{
Description - This function does stuff.
}
\seealso{
\code{\link{getQMat}} for creating a GMRF from a graph
}
