% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-juvdens-functions.R
\name{plotcatch}
\alias{plotcatch}
\title{Title}
\usage{
plotcatch(area, rivs, ctm, gis, add = FALSE)
}
\arguments{
\item{area}{an input parameter}

\item{rivs}{the river network spatial object}

\item{ctm}{the catchment spatial object}

\item{gis}{dont know - sample points perhaps spatial data}

\item{add}{add to an existing plot}
}
\value{
what does it return
}
\description{
Details
}
\details{
Description - This function does stuff.
}
\seealso{
\code{\link{getQMat}} for creating a GMRF from a graph
}
