% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-juvdens-functions.R
\name{plotcatchtile}
\alias{plotcatchtile}
\title{Title}
\usage{
plotcatchtile(area, rivs, gis, tiles)
}
\arguments{
\item{area}{an input parameter}

\item{rivs}{the river network spatial object}

\item{gis}{dont know - sample points perhaps spatial data}

\item{tiles}{the osm tiles}
}
\value{
what does it return
}
\description{
Details
}
\details{
Description - This function does stuff.
}
\seealso{
\code{\link{getQMat}} for creating a GMRF from a graph
}
