% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-juvdens-functions.R
\name{plotrivers}
\alias{plotrivers}
\title{Title}
\usage{
plotrivers(area, rivs, osm = FALSE)
}
\arguments{
\item{area}{an input parameter}

\item{rivs}{the river network spatial object}

\item{osm}{todo}
}
\value{
what does it return
}
\description{
Details
}
\details{
Description - This function does stuff.
}
\seealso{
\code{\link{getQMat}} for creating a GMRF from a graph
}
