% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMRF_models.R
\name{Predict.matrix.gmrf.smooth}
\alias{Predict.matrix.gmrf.smooth}
\title{Estimate capture probabilites from electrofishing data}
\usage{
\method{Predict.matrix}{gmrf.smooth}(object, data)
}
\arguments{
\item{object}{a gmrf smooth object}

\item{data}{a data.frame containing all relavent info}
}
\value{
glm type object
}
\description{
This function uses the marginal likelihood of capture probabilities
to estimate model parameters
}
