% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-juvdens-functions.R
\name{simeQ}
\alias{simeQ}
\title{Simulate from an inhomogenous GMRF}
\usage{
simeQ(eQ, k = 1, tol = 1e-09, rank = NULL)
}
\arguments{
\item{eQ}{an eigen decomposition of a symmetric positive semi-definate matrix corresponding
to an inhomogenous GMRF}

\item{k}{todo}

\item{tol}{tolerance in deciding what is rank deficiency}

\item{rank}{prespecify the rank}
}
\value{
a single draw from with the appropriate covariance structure
}
\description{
Details
}
\details{
This function does stuff.
}
\examples{
#wk <- ctm[ctm $ CATCHMENT \%in\% c(3:13, 15:50),]
## work out neighbours
#wk_nb <- poly2nb(wk, queen = TRUE)
## create a rw2 GMRF precision matrix and simulate some spatial
## structure
#Q <- getQMat(wk_nb)
#x <- simQ(exp(-4)*Q)
#plot(wk, col = grey( pmax(0, pmin(1, (x + 5) / 10))))
}
\seealso{
\code{\link{getQMat}} for creating a GMRF from a graph
}
