% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-juvdens-functions.R
\name{simplify.graph}
\alias{simplify.graph}
\title{Simulate from an inhomogenous GMRF}
\usage{
simplify.graph(g)
}
\arguments{
\item{g}{a graph object}
}
\value{
a simplified graph
}
\description{
Details
}
\details{
This function does stuff.
}
