% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GMRF_models.R
\name{smooth.construct.gmrf.smooth.spec}
\alias{smooth.construct.gmrf.smooth.spec}
\title{Estimate capture probabilites from electrofishing data}
\usage{
\method{smooth.construct}{gmrf.smooth.spec}(object, data, knots)
}
\arguments{
\item{data}{a data.frame containing all relavent info}

\item{formula}{a formula object}
}
\value{
glm type object
}
\description{
This function uses the marginal likelihood of capture probabilities
to estimate model parameters
}

