% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{CLmodel}
\alias{CLmodel}
\alias{CLmodel-package}
\title{Modelling Framework for the Estimation of Salmonid Abundance in Scottish Rivers}
\description{
Modelling Framework for the Estimation of Salmonid Abundance in Scottish Rivers
}

