% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-juvdens-functions.R
\name{plotMyRoute}
\alias{plotMyRoute}
\title{Finds and plots the shortest route between two points
on a river network}
\usage{
plotMyRoute(graph, river)
}
\arguments{
\item{graph}{an input parameter}

\item{river}{todo}
}
\value{
what does it return
}
\description{
Details
}
\details{
Description - This function does stuff.
}
\seealso{
\code{\link{getQMat}} for creating a GMRF from a graph
}
