% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/old-juvdens-functions.R
\name{simplify.graph}
\alias{simplify.graph}
\title{Simulate from an inhomogenous GMRF}
\usage{
simplify.graph(g)
}
\arguments{
\item{Q}{a symmetric positive semi-definate matrix corresponding
to an inhomogenous GMRF}
}
\value{
a single draw from with the appropriate covariance structure
}
\description{
Details
}
\details{
This function does stuff.
}
\seealso{
\code{\link{getQMat}} for creating a GMRF from a graph
}

